break out t test from init test usage
